Component refactorization, dashboard start
