+++ components/content/youtube_project_card.tsx
