+++ components/layout/navbar_dashboard.tsx
@@ -15,7 +15,7 @@ import NextLink from "next/link";
 
 import clsx from "clsx";
 
-import { ThemeSwitch } from "@/components/theme-switch";
+import { ThemeSwitch } from "@/components/layout/theme-switch";
 import { siteConfig } from "@/config/site";
 import { fontMono } from "@/config/fonts";
