+++ components/text/typewriterwrapper.tsx
