+++ src/app/(main)/programming/page.tsx
@@ -1,8 +1,8 @@
-import { Header1Mono, Header2Mono, SubheaderMono } from "@/components/headers";
-import TypewriterWrapper from "@/components/typewriterwrapper";
-import { PageWrapper } from "@/components/pagewrapper";
+import { Header1Mono, Header2Mono, SubheaderMono } from "@/components/text/headers";
+import TypewriterWrapper from "@/components/text/typewriterwrapper";
+import { PageWrapper } from "@/components/page-anim/pagewrapper";
 import { Metadata } from "next";
-import { FadeInScroll } from "@/components/fadeinscroll";
+import { FadeInScroll } from "@/components/page-anim/fadeinscroll";
 import { Button, Chip, Card, CardBody, CardFooter, CardHeader, Divider, Link } from "@heroui/react";
 import { GithubIcon } from "@/components/icons";
 import { getProgrammingProjectsSkillsFull } from "@/src/lib/data/programming";
