+++ src/app/admin/layout.tsx
@@ -1,6 +1,6 @@
 import {ReactNode} from "react";
 import {AuthContextProvider} from "@/components/AuthContext";
-import {NavbarDashboard} from "@/components/navbar_dashboard";
+import {NavbarDashboard} from "@/components/layout/navbar_dashboard";
 
 export const dynamic = "force-dynamic";
