+++ src/app/admin/page.tsx
@@ -1,18 +1,13 @@
-
-import {Header1Mono} from "@/components/headers";
-import {PageWrapper} from "@/components/pagewrapper";
-import {Metadata} from "next";
-
-export const metadata: Metadata = {
-    title: 'Admin Dashboard',
-}
+import {Header1Mono} from "@/components/text/headers";
+import {PageWrapper} from "@/components/page-anim/pagewrapper";
+import Dashboard from "@/components/dashboard/dashboard";
 
 export default function AdminHome() {
     return <PageWrapper>
-        <div className="h-screen flex items-center">
-            <div className="h-min m-auto">
-                <Header1Mono>Dashboard</Header1Mono>
-                <p>Full functionality coming soon!</p>
+        <div className="pt-24">
+            <div className="">
+                <Header1Mono align={"left"} className={"mb-4"}>Dashboard</Header1Mono>
+                <Dashboard />
             </div>
         </div>
     </PageWrapper>
